minor update: complete quest_GetArrayPath
This updates allows the system to find the index within an array by
providing any key and the value that key should be.  This is primarily
designed for finding step indices within the various steps arrays, but
was written generically enough to find any index within any array as
long as the key exists within the array somewhere.